django_auth/requirements.txt
asgiref==3.8.1
Django==3.2.25
django-cors-headers==4.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
python-decouple
mysqlclient
Pillow